feat: Add notes on DJI's Mobile-SDK-Android-V5 project
This commit introduces documentation, `DjiDronesAndroidSdk.md`, that provides notes on DJI's `Mobile-SDK-Android-V5` Android project.
It discusses the strengths and areas for improvement, with a focus on the architectural use of `ViewModel`, `rxJava`, `LiveData`, and `Views`.
It critiques the use of `ViewModel` for global purposes and its injection into `DJIApplication`.
It highlights the injection of the `Application` `Context` into `ViewModel` and suggests refactoring steps.
The commit adds a link to the `DjiDronesAndroidSdk.md` documentation file to the `README.md` file.